ResponserHandler and TwitterStream
changed enum objects to return only the value in the ResponseHandler 
superclass instead of the whole object. Method is no longer 
@staticmethod

Intitialazed variables on extract_data method of StreamResponseHandler 
in order to avoid refering to the variables before initialiazation.

Addded an class variable of sqlHandler to handle the tweet insert